+++ ship_file.json
@@ -1,1 +1,1 @@
-[{"captain": "OwO", "ship_name": "OwO's ship", "cannons": 5, "crew": 5, "armor": 5, "sails": 5, "gold": 900, "win": 0, "loss": 6, "x": 0, "y": 0}, {"captain": "MrBones1102", "cannons": 7, "crew": 10, "armor": 11, "sails": 10, "gold": 454, "win": 0, "loss": 0, "x": 0, "y": 0, "ship_name": "MrBones1102's ship"}, {"captain": "bdavs", "ship_name": "The Revenge", "cannons": 11, "crew": 11, "armor": 6, "sails": 6, "gold": 11006, "win": 6, "loss": 0, "x": 1, "y": 0}, {"captain": "TysonTheCanadian", "cannons": 1, "crew": 1, "armor": 1, "sails": 2, "gold": 15, "win": 0, "loss": 0, "x": 0, "y": 0, "ship_name": "TysonTheCanadian's ship"}]
+[{"captain": "OwO", "ship_name": "OwO's ship", "cannons": 5, "crew": 5, "armor": 5, "sails": 5, "gold": 900, "win": 0, "loss": 9, "x": 0, "y": 0}, {"captain": "MrBones1102", "cannons": 7, "crew": 10, "armor": 11, "sails": 10, "gold": 454, "win": 0, "loss": 0, "x": 0, "y": 0, "ship_name": "MrBones1102's ship"}, {"captain": "bdavs", "ship_name": "The Revenge", "cannons": 11, "crew": 11, "armor": 6, "sails": 6, "gold": 11222, "win": 9, "loss": 0, "x": 1, "y": 0}, {"captain": "TysonTheCanadian", "cannons": 1, "crew": 1, "armor": 1, "sails": 2, "gold": 15, "win": 0, "loss": 0, "x": 0, "y": 0, "ship_name": "TysonTheCanadian's ship"}]
